+++ src/main/java/faang/school/projectservice/controller/ProjectController.java
@@ -1,9 +1,9 @@
 package faang.school.projectservice.controller;
 
-import faang.school.projectservice.dto.client.project.ProjectDto;
+import faang.school.projectservice.dto.project.ProjectDto;
 import faang.school.projectservice.model.ProjectStatus;
 import faang.school.projectservice.model.ProjectVisibility;
-import faang.school.projectservice.service.ProjectService;
+import faang.school.projectservice.service.project.ProjectService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
