--- src/main/java/faang/school/projectservice/mapper/ProjectMapper.java
@@ -1,18 +1,0 @@
-package faang.school.projectservice.mapper;
-
-import faang.school.projectservice.dto.client.project.ProjectDto;
-import faang.school.projectservice.model.Project;
-import org.mapstruct.Mapper;
-import org.mapstruct.Mapping;
-import org.mapstruct.ReportingPolicy;
-
-@Mapper(componentModel = "spring", unmappedTargetPolicy = ReportingPolicy.IGNORE)
-public interface ProjectMapper {
-
-    ProjectDto toDto(Project project);
-
-    @Mapping(target = "createdAt", ignore = true)
-    @Mapping(target = "updatedAt", ignore = true)
-    Project mapToEntity(ProjectDto projectDto);
-    //Juri ne vidit obnovlenija
-}
