+++ src/main/java/faang/school/projectservice/mapper/project/ProjectMapper.java
@@ -6,7 +6,7 @@ import org.mapstruct.Mapper;
 import org.mapstruct.ReportingPolicy;
 
 @Mapper(componentModel = "spring", unmappedSourcePolicy = ReportingPolicy.IGNORE)
-public interface ProjectDtoMapper {
+public interface ProjectMapper {
 
     Project toEntity(ProjectDto projectDto);
