+++ src/test/java/faang/school/projectservice/ProjectControllerTests.java
@@ -3,10 +3,10 @@ package faang.school.projectservice;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import faang.school.projectservice.config.context.UserContext;
 import faang.school.projectservice.controller.ProjectController;
-import faang.school.projectservice.dto.client.project.ProjectDto;
+import faang.school.projectservice.dto.project.ProjectDto;
 import faang.school.projectservice.model.ProjectStatus;
 import faang.school.projectservice.model.ProjectVisibility;
-import faang.school.projectservice.service.ProjectService;
+import faang.school.projectservice.service.project.ProjectService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
